AG-17886 Move demo banner to info icon tooltip
