+++ packages/ag-charts-demos/src/demos/financial/components/SparklineCell.tsx
@@ -40,6 +40,9 @@ function sparklineOptions(container: HTMLElement, points: SparkPoint[], baseline
             key: 'y',
             segments: [{ stop: baseline, stroke: DOWN }],
         },
+        tooltip: {
+            enabled: false,
+        },
     };
 }
